Update patch set 1
Patch Set 1:

I am not sure what you are talking. selinux is there from android 4.3 irrespective of the device. in 4.4 it is in enforcing mode. So we need to patch policy. I havnt tried ntfs patch (I am at work ) Will test and post result with that, it is possible that we might be able remove this completeley of few lines of this patch

Patch-set: 1
Label: Code-Review=0